add initial site superstructure
adds an initial draft of what the structure of the site will
look like.

The basic idea is a landing page with two sections,
"developer profiles" and "resources".
The "developer profiles" will essentially be a table of contents
with links that take readers to flowcharts for each developer profile,
e.g. our single developer that we're iterating on now.

Then from the flowcharts, a reader can click on sections
to read the detailed descriptions we will have in "recipes".

For now the "resources" section just contains the glossary.

- add site/content/_index.md

  with a "developer profiles" section that points to
  the "single developer" flowchart,
  and a separate "resources" section with a link to the glossary

- add ./site./content/flowcharts/ with lonecoder.md
  + this will be the "content section" (Hugo lingo)
    that contains all the graphic depictions of the
    different workflows.

    A site visitor to the top-level

- add ./site/content/glossary.md